api_wordbox/requirements.txt
alembic==1.7.5
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
dataclasses==0.8
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-UUID==0.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.7.7
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
